+++ src/myGui/Operator.java
@@ -2,9 +2,9 @@ package myGui;
 import java.util.Stack;
 
 public abstract class Operator implements Praxeis{
-	protected Stack<Integer> st;
+	protected Stack<Float> st;
 	protected ResultPresenter rp; 
-	public Operator(ResultPresenter rp,Stack<Integer> st2 ){
+	public Operator(ResultPresenter rp,Stack<Float> st2 ){
 		this.st= st2;
 		this.rp=rp;
 	}
